Fix inconsistent usage example in the help message
In the 'Examples' section of the help message, the third usage
example is inconsistent with its description.

    timebox 12      run a 10 minute time box

While the usage example shows how to run a 12 minute time box,
the description mentions that this runs a 10 minute time box.
Correct this usage example to the following.

    timebox 10      run a 10 minute time box

Since this usage message is present in the project README file,
make the same correction in the README file too.