Fix MSG_COMMANDS so that messages appear in dialog box
The value of MSG_COMMANDS variable is a list of commands that must be
invoked to try and display a dialog box at the end of a timebox. The
message is not being displayed because these commands use "$msg" as a
placeholder for the message. At the time when this variable is being
defined, "$msg" evaluates to "", i.e. an empty string.

Fix this issue by escaping the dollar-sign, i.e. use "\$msg" instead of
"$msg", so that it is evaluated later in the eval statement within the
msg function.