Split alltest into unixtest and posixtest
The unixtest target now contains tests for bash, ksh and zsh.
The posixtest target now contains tests for dash and posh.

The behaviour of alltest remains unchanged because it depends on
test, unixtest and posixtest.

This change makes it possible to run the tests on Mac OS X and
non-Debian systems conveniently by running `make unixtest`.