Use POSIX compliant regular expression in tests
In the tests, the regular expression character class "\a" was being used
with the sed command to get rid of the BEL character meant for producing
beeps.  However "\a" is not POSIX compliant. Use "[^[:print:]]" instead
to match non-printable characters. Since the BEL character is not a
printable character, it is matched by this character class.